networkmanager module: support for dns and dhcp
This adds support for specifying the dhcp and dns options to
networkmanager. It just follows upstream's defaults if not specified so
there are no changes in behaviour unless the user requests it.

Especially the ```dns``` part is interesting as it allows us to do split
lookups when connected to a VPN.

Additional stuff

- remove deprecated `hostname`
- make loglevel configurable
- do not log to auditd if auditing is disabled