networkmanager module: support for dns and dhcp options
This adds support for specifying the dhcp and dns options to
networkmanager. It just follows upstream's defaults if not specified.

Especially the ```dns``` part is interesting as it allows us to do split
lookups when connected to a VPN.

Support for ```unbound``` as DNS resolver is currently not supported.

no unbound support for now