homeassistant: [WIP] init at 0.30.2
Also includes a basic nixos module.

This is ridiculously painful due to how dependencies are handled (or I'm
missing something completely).

The basic dependencies for the application itself are declared up front,
but each sub-module has a ```REQUIREMENTS``` array where other python
packages are listed. homeassistant will then try to load those
dependencies if the module is activated and if failing to do so, will
fall back to using pip to install them.

I honestly don't know what the best way to deal with this is. It's a
rabbit hole of dependencies that reminds me of .rpm hell in the old RH
days. Additionally, these sub modules will sometimes require specific
versions which means multiple versions now need to be maintained and the
maintenance cost just shoots up.