freerdp: 20160909 -> 20170201 and dump old freerdp
The former "stable" version of FreeRDP was actually not stable - it just
happened to have a released version while being both buggy and insecure.

The "unstable" branch hasn't seen a release in years, but everybody
should be using this instead and I have been using it as a daily driver
for ages.

This new version works beautifully here.

If/when at some point upstream does a normal release, we can bring back
stable/unstable if needed.

As I am quite dependent on FreeRDP working properly, I will be commit to
keeping this updated.

The commit used for this release follows the Arch Linux release.